Create Program to perform linear search and binary search in a array
First of all this program will ask the size of the array. Then the user will have to input the elements of the array. Then it will ask you to choose between the linear search and binary search, then enter the element u want to search and then its position from starting will be shown.